+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 
 - Add support for querying Orders by `metadata`
+- Added `transaction_value_eth_gwei` as an alternative way to compute transaction level emissions for ethereum
 
 ## [1.8.0] - 2021-07-20